--- app/src/main/java/com/example/jonathan/sectionsapp/model/properties/Header.kt
@@ -1,3 +1,0 @@
-package com.example.jonathan.sectionsapp.model.properties
-
-interface Header: Element
