--- app/src/main/java/com/example/jonathan/sectionsapp/model/properties/Headline.kt
@@ -1,5 +1,0 @@
-package com.example.jonathan.sectionsapp.model.properties
-
-data class Headline(
-    val text: Text
-): Header
