--- app/src/main/java/com/example/jonathan/sectionsapp/model/properties/Shape.kt
@@ -1,5 +1,0 @@
-package com.example.jonathan.sectionsapp.model.properties
-
-enum class Shape {
-    CIRCLE, SQUARE, RECTANGLE
-}
