--- app/src/main/java/com/example/jonathan/sectionsapp/model/section/GridSection.kt
@@ -1,11 +1,0 @@
-package com.example.jonathan.sectionsapp.model.section
-
-import com.example.jonathan.sectionsapp.model.item.SectionItem
-import com.example.jonathan.sectionsapp.model.properties.Header
-
-data class GridSection(
-    override val id: Int,
-    override val header: Header?,
-    override val items: List<SectionItem>,
-    val columnsCount: Int
-): Section
