+++ domain/src/main/java/com/example/jonathan/domain/model/properties/DiscoverHeader.kt
@@ -1,4 +1,4 @@
-package com.example.jonathan.sectionsapp.model.properties
+package com.example.jonathan.domain.model.properties
 
 data class DiscoverHeader(
     val cover: Cover,
