+++ domain/src/main/java/com/example/jonathan/domain/model/properties/Element.kt
@@ -1,0 +1,3 @@
+package com.example.jonathan.domain.model.properties
+
+interface Element
