+++ domain/src/main/java/com/example/jonathan/domain/model/properties/Headline.kt
@@ -1,0 +1,5 @@
+package com.example.jonathan.domain.model.properties
+
+data class Headline(
+    val text: Text
+): Header
