+++ domain/src/main/java/com/example/jonathan/domain/model/properties/Shape.kt
@@ -1,0 +1,5 @@
+package com.example.jonathan.domain.model.properties
+
+enum class Shape {
+    CIRCLE, SQUARE, RECTANGLE
+}
