+++ domain/src/main/java/com/example/jonathan/domain/model/properties/SizeType.kt
@@ -1,0 +1,5 @@
+package com.example.jonathan.domain.model.properties
+
+enum class SizeType {
+    SMALL, MEDIUM, LARGE
+}
