+++ domain/src/main/java/com/example/jonathan/domain/model/section/GridSection.kt
@@ -1,0 +1,11 @@
+package com.example.jonathan.domain.model.section
+
+import com.example.jonathan.domain.model.item.SectionItem
+import com.example.jonathan.domain.model.properties.Header
+
+data class GridSection(
+    override val id: Int,
+    override val header: Header?,
+    override val items: List<SectionItem>,
+    val columnsCount: Int
+): Section
