+++ domain/src/main/java/com/example/jonathan/domain/model/section/Section.kt
@@ -1,0 +1,10 @@
+package com.example.jonathan.domain.model.section
+
+import com.example.jonathan.domain.model.item.SectionItem
+import com.example.jonathan.domain.model.properties.Header
+
+interface Section {
+    val id: Int
+    val header: Header?
+    val items: List<SectionItem>
+}
